Updates to wristband classes, scaffold wristband UI fragments
Added WristbandPrinterFragment to display print status and provide
print button. UI is not polished at this point, but enough to support
development.

Added WristbandPrinterListFragment to display up to 3 configured
wristband printers (using instances of WristbandPrinterFragment
as child fragment).

Refactored code in the wristband classes. Changes to connect and
disconnect to allow them to be called with a job to check if
printer is reachable on the network.

Tested:

created a shim Activity in Register app that uses WristbandPrinterListFragment
and configured 3 printers. Don't currently have a printer but it did
correctly report in the UI the printer specifics and that it was unable
to connect to any of them.