add wristband job status and connectivity threads
Added thread to WristbandPrinterFragment that

- sleep 5 seconds
- call down to wristband printer to check reachability of printer at configured ip/port
- update on on UI thread the reachability status

Similar thread added to do checks of print job status (this is experimental, it may
be that the registered callback for print job status is a better way to do this)

Threads current check on start then poll every 5 seconds.

Test

Verified without any printers on the network, print buttons are disabled and that
the printer reports as not connected.

Testing needed

configure printer 1 in config UI for printer ip/port
turn on and off the printer (or unplug the network cable) and see that the UI
responds every 5 seconds with correct state.

repeat above for printers 2 and 3, make sure correct printer in UI is updated.